Update search result display
