+++ src/components/param_inputs/RosTypeParam.vue
@@ -32,8 +32,8 @@
 import { useEditNodeStore } from '@/stores/edit_node';
 import { useEditorStore } from '@/stores/editor';
 import { useMessasgeStore } from '@/stores/message';
-import { RosActionType_Name, RosServiceType_Name, RosTopicType_Name, type RosName, type RosType } from '@/types/python_types';
-import type { Channel, Message, ParamData } from '@/types/types';
+import type { RosType } from '@/types/python_types';
+import type { ParamData } from '@/types/types';
 import Fuse from 'fuse.js';
 import { computed, ref } from 'vue';
 
@@ -48,38 +48,20 @@ const props = defineProps<{
   type: 'topic' | 'service' | 'action'
 }>()
 
-let search_results = ref<Channel[]>([])
+let search_results = ref<string[]>([])
 
 const param = computed<ParamData | undefined>(() =>
   edit_node_store.new_node_options.find((x) => x.key === props.data_key)
 )
 
-// Find a type parameter to reference when searching and set when selecting
-const type_param = computed<ParamData | undefined>(() => {
-  let type_param_name: string
+const search_fuse = computed<Fuse<string> | undefined>(() => {
   switch (props.type) {
     case 'topic':
-      type_param_name = RosTopicType_Name
-      break
+      return messages_store.ros_topic_type_fuse
     case 'service':
-      type_param_name = RosServiceType_Name
-      break
+      return messages_store.ros_service_type_fuse
     case 'action':
-      type_param_name = RosActionType_Name
-    default:
-      return undefined
-  }
-  return edit_node_store.new_node_options.find((x) => x.value.type === type_param_name)
-})
-
-const search_fuse = computed<Fuse<Channel> | undefined>(() => {
-  switch (props.type) {
-    case 'topic':
-      return messages_store.ros_topic_name_fuse
-    case 'service':
-      return messages_store.ros_service_name_fuse
-    case 'action':
-      return messages_store.ros_action_name_fuse
+      return messages_store.ros_action_type_fuse
     default:
       return undefined
   }
@@ -99,23 +81,11 @@ function onInput(event: Event) {
   }
 
   const target = event.target as HTMLInputElement
-  let new_name = target.value || ''
+  let new_type_name = target.value || ''
+  const results = search_fuse.value.search(new_type_name)
+  search_results.value = results.map((x) => x.item)
 
-  let results
-  if (type_param.value === undefined) {
-    results = search_fuse.value.search({name: new_name})
-  } else {
-    const type = (type_param.value.value.value as RosType).type_str
-    if (type === '') {
-      results = search_fuse.value.search({name: new_name})
-    } else {
-      results = search_fuse.value.search({name: new_name, type: type})
-    }
-  }
-
-  search_results.value = results.slice(0, 3).map((x) => x.item)
-
-  setValue(new_name)
+  setValue(new_type_name)
 }
 
 function setValue(new_value: string) {
@@ -124,27 +94,14 @@ function setValue(new_value: string) {
     return
   }
 
-  let name_obj = param.value.value.value as RosName
-  name_obj.name = new_value
-
-  edit_node_store.updateParamValue(props.category, props.data_key, name_obj)
-}
-
-function setType(new_value: string) {
-  if (type_param.value === undefined) {
-    // No type param to set
-    return
-  }
-
-  let type_obj = type_param.value.value.value as RosType
+  let type_obj = param.value.value.value as RosType
   type_obj.type_str = new_value
 
-  edit_node_store.updateParamValue(props.category, type_param.value.key, type_obj)
+  edit_node_store.updateParamValue(props.category, props.data_key, type_obj)
 }
 
-function selectSearchResult(search_result: Channel) {
-  setValue(search_result.name)
-  setType(search_result.type)
+function selectSearchResult(search_result: string) {
+  setValue(search_result)
   releaseDropdown()
 }
 
@@ -173,7 +130,7 @@ function releaseDropdown() {
       <input
         type="text"
         class="form-control mt-2"
-        :value="(param.value.value as RosName).name"
+        :value="(param.value.value as RosType).type_str"
         :disabled="editor_store.selected_subtree.is_subtree"
         @input="onInput"
         @focus="focusInput"
@@ -189,22 +146,21 @@ function releaseDropdown() {
     </label>
     <div class="mb-2 search-results">
       <div
-        class="list-group"
+        class="list-group rounded-top-0"
         :class="{ 'd-none': hide_results && !keep_results }"
         @mouseenter="forceDropdown"
         @mouseleave="releaseDropdown"
       >
         <div
           v-for="result in search_results"
-          :key="result.name"
+          :key="result"
           class="list-group-item search-result"
           tabindex="0"
           @click="() => selectSearchResult(result)"
           @keyup.enter="() => selectSearchResult(result)"
           @keyup.esc="releaseDropdown"
         >
-          {{ result.name }}<br>
-          <small>{{ result.type }}</small>
+          {{ result }}
         </div>
       </div>
     </div>
@@ -213,15 +169,5 @@ function releaseDropdown() {
 </template>
 
 <style scoped lang="scss">
-.search-results {
-  padding-left: 10px;
-}
 
-.search-result:hover {
-  background-color: #007bff;
-}
-
-.search-result-highlighted {
-  background-color: #007bff;
-}
 </style>
